requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
python-multipart
openai
google-generativeai